+++ src/XPath/NodeListImplSurrogate.hpp
@@ -88,7 +88,7 @@ public:
 	// These interfaces are inherited from NodeListImpl...
 
     virtual NodeImpl*
-	item(int index);
+	item(unsigned long	index);
 
     virtual int
 	getLength();
